Implementa funcionalidades de relatórios e melhorias no módulo de obras
- Adiciona funções para imprimir colaborações e listar todas as colaborações no módulo de colaborações.
- Melhora o módulo de obras com novos menus e opções para listar, cadastrar, alterar e excluir obras.
- Implementa relatórios para exibir obras por artista e por tipo, além de colaborações em um período específico.
- Atualiza a persistência para suportar valores em centavos como long long.
- Refatora funções de leitura e impressão de valores para lidar com long long.
- Remove arquivos binários antigos e adiciona novos arquivos de persistência.
- Gera relatórios em arquivos de texto para facilitar a visualização dos dados.

+++ estruturas.h
@@ -71,7 +71,7 @@ typedef struct
 	int id; // ID única para cada obra.
 	int anoCriacao;
 	
-	int valorCentavos; // Valor em centavos para evitar erro de conversão float.
+	long long valorCentavos; // Valor em centavos para evitar erro de conversão float, precisei declara como long porque obras de arte tem valores muito altos.
 
 	char titulo[TAM_TEXTO_MEDIO];
 	char tipo[TAM_TEXTO_PEQUENO];
